refactor: proxima app to own file in view
